+++ swap/tests/e2e.rs
@@ -1,19 +1,17 @@
+use crate::testutils::{init_alice, init_bob};
 use bitcoin_harness::Bitcoind;
 use futures::future::try_join;
 use libp2p::Multiaddr;
 use monero_harness::Monero;
 use rand::rngs::OsRng;
-use std::sync::Arc;
-use swap::{
-    alice, alice::swap::AliceState, bob, bob::swap::BobState, network::transport::build,
-    storage::Database, SwapAmounts,
-};
+use swap::{alice, alice::swap::AliceState, bob, bob::swap::BobState, storage::Database};
 use tempfile::tempdir;
 use testcontainers::clients::Cli;
-use tracing_core::dispatcher::DefaultGuard;
-use tracing_log::LogTracer;
+use testutils::init_tracing;
 use uuid::Uuid;
-use xmr_btc::{alice::State0, bitcoin, config::Config, cross_curve_dleq};
+use xmr_btc::{bitcoin, config::Config};
+
+pub mod testutils;
 
 /// Run the following tests with RUST_MIN_STACK=10000000
 
@@ -48,8 +46,8 @@ async fn happy_path() {
 
     let (
         alice_state,
-        mut alice_swarm_driver,
-        alice_swarm_handle,
+        mut alice_event_loop,
+        alice_event_loop_handle,
         alice_btc_wallet,
         alice_xmr_wallet,
     ) = init_alice(
@@ -64,7 +62,7 @@ async fn happy_path() {
     )
     .await;
 
-    let (bob_state, bob_swarm_driver, bob_swarm_handle, bob_btc_wallet, bob_xmr_wallet, bob_db) =
+    let (bob_state, bob_event_loop, bob_event_loop_handle, bob_btc_wallet, bob_xmr_wallet, bob_db) =
         init_bob(
             alice_multiaddr,
             &bitcoind,
@@ -82,7 +80,7 @@ async fn happy_path() {
 
     let alice_swap_fut = alice::swap::swap(
         alice_state,
-        alice_swarm_handle,
+        alice_event_loop_handle,
         alice_btc_wallet.clone(),
         alice_xmr_wallet.clone(),
         config,
@@ -90,11 +88,11 @@ async fn happy_path() {
         alice_db,
     );
 
-    let _alice_swarm_fut = tokio::spawn(async move { alice_swarm_driver.run().await });
+    let _alice_swarm_fut = tokio::spawn(async move { alice_event_loop.run().await });
 
     let bob_swap_fut = bob::swap::swap(
         bob_state,
-        bob_swarm_handle,
+        bob_event_loop_handle,
         bob_db,
         bob_btc_wallet.clone(),
         bob_xmr_wallet.clone(),
@@ -102,7 +100,7 @@ async fn happy_path() {
         Uuid::new_v4(),
     );
 
-    let _bob_swarm_fut = tokio::spawn(async move { bob_swarm_driver.run().await });
+    let _bob_swarm_fut = tokio::spawn(async move { bob_event_loop.run().await });
 
     try_join(alice_swap_fut, bob_swap_fut).await.unwrap();
 
@@ -154,20 +152,25 @@ async fn alice_punishes_if_bob_never_acts_after_fund() {
 
     let config = Config::regtest();
 
-    let (alice_state, mut alice_swarm, alice_swarm_handle, alice_btc_wallet, alice_xmr_wallet) =
-        init_alice(
-            &bitcoind,
-            &monero,
-            btc_to_swap,
-            alice_btc_starting_balance,
-            xmr_to_swap,
-            alice_xmr_starting_balance,
-            alice_multiaddr.clone(),
-            config,
-        )
-        .await;
+    let (
+        alice_state,
+        mut alice_event_loop,
+        alice_event_loop_handle,
+        alice_btc_wallet,
+        alice_xmr_wallet,
+    ) = init_alice(
+        &bitcoind,
+        &monero,
+        btc_to_swap,
+        alice_btc_starting_balance,
+        xmr_to_swap,
+        alice_xmr_starting_balance,
+        alice_multiaddr.clone(),
+        config,
+    )
+    .await;
 
-    let (bob_state, bob_swarm_driver, bob_swarm_handle, bob_btc_wallet, bob_xmr_wallet, bob_db) =
+    let (bob_state, bob_event_loop, bob_event_loop_handle, bob_btc_wallet, bob_xmr_wallet, bob_db) =
         init_bob(
             alice_multiaddr,
             &bitcoind,
@@ -183,7 +186,7 @@ async fn alice_punishes_if_bob_never_acts_after_fund() {
     let bob_btc_locked_fut = bob::swap::run_until(
         bob_state,
         bob::swap::is_btc_locked,
-        bob_swarm_handle,
+        bob_event_loop_handle,
         bob_db,
         bob_btc_wallet.clone(),
         bob_xmr_wallet.clone(),
@@ -191,14 +194,14 @@ async fn alice_punishes_if_bob_never_acts_after_fund() {
         Uuid::new_v4(),
     );
 
-    let _bob_swarm_fut = tokio::spawn(async move { bob_swarm_driver.run().await });
+    let _bob_swarm_fut = tokio::spawn(async move { bob_event_loop.run().await });
 
     let alice_db_datadir = tempdir().unwrap();
     let alice_db = Database::open(alice_db_datadir.path()).unwrap();
 
     let alice_fut = alice::swap::swap(
         alice_state,
-        alice_swarm_handle,
+        alice_event_loop_handle,
         alice_btc_wallet.clone(),
         alice_xmr_wallet.clone(),
         Config::regtest(),
@@ -206,7 +209,7 @@ async fn alice_punishes_if_bob_never_acts_after_fund() {
         alice_db,
     );
 
-    let _alice_swarm_fut = tokio::spawn(async move { alice_swarm.run().await });
+    let _alice_swarm_fut = tokio::spawn(async move { alice_event_loop.run().await });
 
     // Wait until alice has locked xmr and bob has locked btc
     let ((alice_state, _), bob_state) = try_join(alice_fut, bob_btc_locked_fut).await.unwrap();
@@ -314,6 +317,10 @@ async fn both_refund() {
 
     tokio::spawn(async move { bob_swarm_driver.run().await });
 
+    let alice_swap_id = Uuid::new_v4();
+    let alice_db_datadir = tempdir().unwrap();
+    let alice_db = Database::open(alice_db_datadir.path()).unwrap();
+
     let alice_xmr_locked_fut = alice::swap::run_until(
         alice_state,
         alice::swap::is_xmr_locked,
@@ -321,8 +328,8 @@ async fn both_refund() {
         alice_btc_wallet.clone(),
         alice_xmr_wallet.clone(),
         Config::regtest(),
-        todo!(),
-        todo!(),
+        alice_swap_id,
+        alice_db,
     );
 
     tokio::spawn(async move { alice_swarm_driver.run().await });
@@ -337,14 +344,16 @@ async fn both_refund() {
         panic!("Bob in unexpected state");
     };
 
+    let alice_db = Database::open(alice_db_datadir.path()).unwrap();
+
     let (alice_state, _) = alice::swap::swap(
         alice_state,
         alice_swarm_handle,
         alice_btc_wallet.clone(),
         alice_xmr_wallet.clone(),
         Config::regtest(),
-        todo!(),
-        todo!(),
+        alice_swap_id,
+        alice_db,
     )
     .await
     .unwrap();
@@ -383,190 +392,4 @@ async fn both_refund() {
     bob_xmr_wallet.as_ref().0.refresh().await.unwrap();
     let xmr_bob_final = bob_xmr_wallet.as_ref().get_balance().await.unwrap();
     assert_eq!(xmr_bob_final, bob_xmr_starting_balance);
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn init_alice(
-    bitcoind: &Bitcoind<'_>,
-    monero: &Monero,
-    btc_to_swap: bitcoin::Amount,
-    _btc_starting_balance: bitcoin::Amount,
-    xmr_to_swap: xmr_btc::monero::Amount,
-    xmr_starting_balance: xmr_btc::monero::Amount,
-    listen: Multiaddr,
-    config: Config,
-) -> (
-    AliceState,
-    alice::event_loop::EventLoop,
-    alice::event_loop::EventLoopHandle,
-    Arc<swap::bitcoin::Wallet>,
-    Arc<swap::monero::Wallet>,
-) {
-    monero
-        .init(vec![("alice", xmr_starting_balance.as_piconero())])
-        .await
-        .unwrap();
-
-    let alice_xmr_wallet = Arc::new(swap::monero::Wallet(
-        monero.wallet("alice").unwrap().client(),
-    ));
-
-    let alice_btc_wallet = Arc::new(
-        swap::bitcoin::Wallet::new("alice", bitcoind.node_url.clone(), config.bitcoin_network)
-            .await
-            .unwrap(),
-    );
-
-    let amounts = SwapAmounts {
-        btc: btc_to_swap,
-        xmr: xmr_to_swap,
-    };
-
-    let rng = &mut OsRng;
-    let (alice_state, alice_behaviour) = {
-        let a = bitcoin::SecretKey::new_random(rng);
-        let s_a = cross_curve_dleq::Scalar::random(rng);
-        let v_a = xmr_btc::monero::PrivateViewKey::new_random(rng);
-        let redeem_address = alice_btc_wallet.as_ref().new_address().await.unwrap();
-        let punish_address = redeem_address.clone();
-        let state0 = State0::new(
-            a,
-            s_a,
-            v_a,
-            amounts.btc,
-            amounts.xmr,
-            config.bitcoin_refund_timelock,
-            config.bitcoin_punish_timelock,
-            redeem_address,
-            punish_address,
-        );
-
-        (
-            AliceState::Started {
-                amounts,
-                state0: state0.clone(),
-            },
-            alice::Behaviour::new(state0),
-        )
-    };
-
-    let alice_transport = build(alice_behaviour.identity()).unwrap();
-
-    let (swarm_driver, handle) =
-        alice::event_loop::EventLoop::new(alice_transport, alice_behaviour, listen).unwrap();
-
-    (
-        alice_state,
-        swarm_driver,
-        handle,
-        alice_btc_wallet,
-        alice_xmr_wallet,
-    )
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn init_bob(
-    alice_multiaddr: Multiaddr,
-    bitcoind: &Bitcoind<'_>,
-    monero: &Monero,
-    btc_to_swap: bitcoin::Amount,
-    btc_starting_balance: bitcoin::Amount,
-    xmr_to_swap: xmr_btc::monero::Amount,
-    xmr_stating_balance: xmr_btc::monero::Amount,
-    config: Config,
-) -> (
-    BobState,
-    bob::event_loop::EventLoop,
-    bob::event_loop::EventLoopHandle,
-    Arc<swap::bitcoin::Wallet>,
-    Arc<swap::monero::Wallet>,
-    Database,
-) {
-    let bob_btc_wallet = Arc::new(
-        swap::bitcoin::Wallet::new("bob", bitcoind.node_url.clone(), config.bitcoin_network)
-            .await
-            .unwrap(),
-    );
-    bitcoind
-        .mint(
-            bob_btc_wallet.inner.new_address().await.unwrap(),
-            btc_starting_balance,
-        )
-        .await
-        .unwrap();
-
-    monero
-        .init(vec![("bob", xmr_stating_balance.as_piconero())])
-        .await
-        .unwrap();
-
-    let bob_xmr_wallet = Arc::new(swap::monero::Wallet(monero.wallet("bob").unwrap().client()));
-
-    let amounts = SwapAmounts {
-        btc: btc_to_swap,
-        xmr: xmr_to_swap,
-    };
-
-    let bob_db_dir = tempdir().unwrap();
-    let bob_db = Database::open(bob_db_dir.path()).unwrap();
-    let bob_behaviour = bob::Behaviour::default();
-    let bob_transport = build(bob_behaviour.identity()).unwrap();
-
-    let refund_address = bob_btc_wallet.new_address().await.unwrap();
-    let state0 = xmr_btc::bob::State0::new(
-        &mut OsRng,
-        btc_to_swap,
-        xmr_to_swap,
-        config.bitcoin_refund_timelock,
-        config.bitcoin_punish_timelock,
-        refund_address,
-    );
-    let bob_state = BobState::Started {
-        state0,
-        amounts,
-        addr: alice_multiaddr,
-    };
-
-    let (swarm_driver, swarm_handle) =
-        bob::event_loop::EventLoop::new(bob_transport, bob_behaviour).unwrap();
-
-    (
-        bob_state,
-        swarm_driver,
-        swarm_handle,
-        bob_btc_wallet,
-        bob_xmr_wallet,
-        bob_db,
-    )
-}
-
-/// Utility function to initialize logging in the test environment.
-/// Note that you have to keep the `_guard` in scope after calling in test:
-///
-/// ```rust
-/// let _guard = init_tracing();
-/// ```
-fn init_tracing() -> DefaultGuard {
-    // converts all log records into tracing events
-    // Note: Make sure to initialize without unwrapping, otherwise this causes
-    // trouble when running multiple tests.
-    let _ = LogTracer::init();
-
-    let global_filter = tracing::Level::WARN;
-    let swap_filter = tracing::Level::DEBUG;
-    let xmr_btc_filter = tracing::Level::DEBUG;
-    let monero_harness_filter = tracing::Level::INFO;
-    let bitcoin_harness_filter = tracing::Level::INFO;
-
-    use tracing_subscriber::util::SubscriberInitExt as _;
-    tracing_subscriber::fmt()
-        .with_env_filter(format!(
-            "{},swap={},xmr-btc={},monero_harness={},bitcoin_harness={}",
-            global_filter,
-            swap_filter,
-            xmr_btc_filter,
-            monero_harness_filter,
-            bitcoin_harness_filter,
-        ))
-        .set_default()
 }
